+++ azure_terminal/module_tests/math_module_test.py
@@ -1,6 +1,8 @@
-# Importing
-from azure_terminal.modules.math_module import execute_math
+# Python Importing
 import random
+
+# Local Importing
+from azure_terminal.modules.math_module import execute_math
 
 def random_operator_generator():
     number: int = random.randint(0,4)