+++ system_library/ChromaFlux/ansi.py
@@ -1,9 +1,3 @@
-# Copyright Jonathan Hartley 2013. BSD 3-Clause license, see LICENSE file.
-'''
-This module generates ANSI character codes to printing colors to terminals.
-See: http://en.wikipedia.org/wiki/ANSI_escape_code
-'''
-
 CSI = '\033['
 OSC = '\033]'
 BEL = '\a'
@@ -58,14 +52,14 @@ class AnsiColour(AnsiCodes):
     RESET           = 39
 
     # These are fairly well supported, but not part of the standard.
-    LIGHTBLACK_EX   = 90
-    LIGHTRED_EX     = 91
-    LIGHTGREEN_EX   = 92
-    LIGHTYELLOW_EX  = 93
-    LIGHTBLUE_EX    = 94
-    LIGHTMAGENTA_EX = 95
-    LIGHTCYAN_EX    = 96
-    LIGHTWHITE_EX   = 97
+    LIGHTBLACK   = 90
+    LIGHTRED     = 91
+    LIGHTGREEN   = 92
+    LIGHTYELLOW  = 93
+    LIGHTBLUE    = 94
+    LIGHTMAGENTA = 95
+    LIGHTCYAN    = 96
+    LIGHTWHITE   = 97
 
 
 class AnsiBack(AnsiCodes):
@@ -80,14 +74,14 @@ class AnsiBack(AnsiCodes):
     RESET           = 49
 
     # These are fairly well supported, but not part of the standard.
-    LIGHTBLACK_EX   = 100
-    LIGHTRED_EX     = 101
-    LIGHTGREEN_EX   = 102
-    LIGHTYELLOW_EX  = 103
-    LIGHTBLUE_EX    = 104
-    LIGHTMAGENTA_EX = 105
-    LIGHTCYAN_EX    = 106
-    LIGHTWHITE_EX   = 107
+    LIGHTBLACK   = 100
+    LIGHTRED     = 101
+    LIGHTGREEN   = 102
+    LIGHTYELLOW  = 103
+    LIGHTBLUE    = 104
+    LIGHTMAGENTA = 105
+    LIGHTCYAN    = 106
+    LIGHTWHITE   = 107
 
 
 class AnsiStyle(AnsiCodes):